Fix QR decoder unit test literals; release alpha.28
The noise fixture used 128 as a Byte, which is out of range; use 128.toByte() and make the grayscale branch explicit. The main APK already compiled in alpha.27; this fixes the unit-test compile step.

+++ apps/mobile/android/app/src/test/java/io/github/sayach/dshmobile/QrDecoderTest.kt
@@ -18,7 +18,7 @@ class QrDecoderTest {
 
     @Test
     fun returnsNullForUnstructuredNoise() {
-        val noise = ByteArray(100 * 100) { 128 }
+        val noise = ByteArray(100 * 100) { 128.toByte() }
         assertNull(QrDecoder.decodeNv21(noise, 100, 100))
     }
 
@@ -31,7 +31,7 @@ class QrDecoderTest {
         val data = ByteArray(size * size)
         for (y in 0 until size) {
             for (x in 0 until size) {
-                data[y * size + x] = if (matrix.get(x, y)) 0 else 0xff.toByte()
+                data[y * size + x] = if (matrix.get(x, y)) 0.toByte() else 0xff.toByte()
             }
         }
         return data
